requirements.txt
# Core dependencies
psycopg2-binary>=2.9.5
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.7

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0

# Data processing
scipy>=1.10.0
matplotlib>=3.7.0

# Documentation
sphinx>=6.1.3
sphinx-rtd-theme>=1.2.0